# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DOCUMENTATION",
        "version": "0.1"
    },
    "paths": {
        "/admin/login": {
            "get": {
                "summary": "showLoginForm()",
                "description": "Show the application's login form.",
                "responses": {
                    "200": {
                        "description": "show Login Form"
                    }
                }
            },
            "post": {
                "summary": "login(Request $request)",
                "description": " Handle a login request to the application.",
                "parameters": [
                    {
                        "name": "\\Illuminate\\Http\\Request  $request",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}